refactor(ui): route workspace dispatch to per-domain handlers
CODE-4: the ~848-line flat `dispatch` match now delegates each command
family to a sibling module (documents, connections, charts_dashboards,
query, scripts, navigation, settings, audit), mirroring the existing
actions/ split. mod.rs keeps a thin dispatcher. Pure code motion; same
commands, order and error paths.
