+++ crates/dbflux_ui/src/ui/views/workspace/dispatch/scripts.rs
@@ -1,0 +1,18 @@
+use super::*;
+
+impl Workspace {
+    pub(super) fn dispatch_scripts(
+        &mut self,
+        cmd: Command,
+        window: &mut Window,
+        cx: &mut Context<Self>,
+    ) -> Option<bool> {
+        match cmd {
+            Command::OpenScriptFile => {
+                self.open_script_file(window, cx);
+                Some(true)
+            }
+            _ => None,
+        }
+    }
+}
